notification_service/requirements.txt
amqp==5.2.0
Django==5.0.2
django-cors-headers==4.3.1
django-otp==1.3.0
django-phonenumber-field==7.3.0
django-redis==5.4.0
djangorestframework==3.14.0
gunicorn==21.2.0
pyotp==2.9.0
python-decouple==3.8
pytz==2024.1
redis==5.0.2
pika

